requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
gunicorn>=23.0.0
flask-wtf>=1.2.2
jinja2>=3.1.6
werkzeug>=3.1.3
python-dotenv>=1.0.0
flask-sqlalchemy>=3.1.1
PyMySQL>=1.1.0
mysql-connector-python>=8.1.0
Pillow>=11.2.1
